CameraViewController dies (UI modernization PR 12)
The file that opened this series at 1,729 lines is deleted. Its remains
split in two: CameraRig (366 lines, non-UI) owns the capture stack, the
screen's view model and the countdown chime/torch, and is the production
CameraControlling conformer — the object handed to the actor system is no
longer a view controller. CameraHostController (180 lines, a
UIHostingController<CameraScreenView> subclass) is the disposable shell:
navigation chrome, permissions UI, modal presentation, rotation forwarding,
and BecomeCamera/UnbecomeCamera at lifecycle edges, reaching the rig
through four closure seams.

Retired along the way: the objc_setAssociatedObject dance for the mic
prompt (the shell has a stored property) and exitCamera's direct
navigationController reach (now the onExit seam).

WatchRemoteCameraController embeds the new pair; the scanner pushes the
shell. This closes out the UI modernization plan — every screen is SwiftUI
behind a thin shell, no actor-facing object is a UIKit type — and
pre-positions MODERNIZATION.md Phase 4 (SessionCoordinator talks to a
plain object).

393/393 tests green on the first run after the deletion.
